refactor: split user-service into smaller services
